% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getInsertSql}
\alias{getInsertSql}
\title{Get insert SQL}
\usage{
getInsertSql(connectionDetails)
}
\value{
list of sql statements
}
\description{
Get insert SQL
}
\details{
Returns insert Sql for native DB.
}

