% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getTestSql}
\alias{getTestSql}
\title{Get testing SQL}
\usage{
getTestSql(connectionDetails)
}
\value{
list of sql statements
}
\description{
Get testing SQL
}
\details{
Returns testing Sql for the CDM DB.
}

